Update ocean.md
I have changed the order of the options, on what I perceive to be a complexity axis. My notion of what this is may be incorrect, but I intended to go from simplest to most complex.

Also, in the SlabOcean energy balance you have listed the sensible heat flux as S, but it is more common to use H, as S is normally dedicated to Salinity. Would it hurt anything to change to H for sensible heat flux?