ThermodynamicSeaIce testing (#814)
* ThermodynamicSeaIce testing

* update changelog

* sea ice tests added

* Ocean and sea ice documentation

* test file path corrected

* increase test tolerance for grid interpolation

* ocean/sea ice generator calls in tests corrected

* docs examples corrected

* more tests and docs corrected

* add isfinite check for sst used in sea ice kernel

* Update sea_ice.md

Fixed a couple of typos

* Update ocean.md

I have changed the order of the options, on what I perceive to be a complexity axis. My notion of what this is may be incorrect, but I intended to go from simplest to most complex.

Also, in the SlabOcean energy balance you have listed the sensible heat flux as S, but it is more common to use H, as S is normally dedicated to Salinity. Would it hurt anything to change to H for sensible heat flux?

* ocean docs code block corrected

---------

Co-authored-by: Pier Luigi Vidale <108409300+pierluigividale@users.noreply.github.com>