Snow fall from large-scale condensation (#817)
* Update diagnostic_variables.jl

Added diagnostic variables for snow

* Update large_scale_condensation.jl

Added rule to compute large scale snow depending on lowest level temperature

* Update large_scale_condensation.jl

Expanded rule for large scale snow

* Update convection.jl

Added rules for convective snow

* Update precipitation.jl

Added diagnostics for snow

* Update large_scale_condensation.jl

Changed test for snow vs precipitation to avoid IF and to deal with levels.

* Update convection.jl

Changed the convective rain and snow decision with no IF statement and scanning all layers.

* Update convection.jl

Attempted to include same code for convective rain and snow

* Update precipitation.jl

Removed duplications in function, abstract variables, modified to single AbstractRateOutputVariable

* Update column_variables.jl

Added diagnostic for all variants of snow variables

* Update precipitation.jl

Eliminated stupid line with "------"

* Move snow/rain decision into vertical loops

* update changelog

* unpack freezing_threshold

* rename to correct LargeScaleSnowOutput

* reset snow in column

* cap convection precip at zero

* Update thermodynamics.jl

Added CC constants for ice species

* Update large_scale_condensation.jl

Added the latent heat exchange associated with freezing precipitation into snow.

* Update large_scale_condensation.jl

Simplified the test for enthalpy calculation when snow forms.
Added a test to melt snow if any layers below the layer of snow formation are above a melting threshold. This is quite conservative, in that a single warm layer below k is enough to destroy snow at k.

* Update large_scale_condensation.jl

Changed the logic for melting snow, as before it was not possible to assign enthalpy to the layer that actually melts the snow.

* Update large_scale_condensation.jl

Fixed units and inserted a check for available snow to melt

* Update large_scale_condensation.jl

Forgot to end the IF loop

* Update large_scale_condensation.jl

Many typos and unit inconsistencies

* Update large_scale_condensation.jl

Had forgotten to change snow for snow_flux near the end

* Update large_scale_condensation.jl

There was a conceptual error with where/how snow_flux is used to update the column snow.

* Update convection.jl

Implemented melting of snow into the BM scheme.
There may be some conceptual errors, someone needs to check this.

* Update convection.jl

There was a missing definition for unit conversions, moved a statement up before the snow melting block.

* Update convection.jl

Fixed an error with the unit corrections, as well as removed calculation from the loop, as it is not necessary to compute the correction for each layer.

* Reevaporation, sublimation and snow fall

* bugs + comment out snow from convection

* latent heat release was overwritten

* avoid negative precipitation

* snow removed from convection

* Don't output snow convection variable

* some bugs and docs

* rain/snow fluxes in m/s

* docs updated

* large-scale condensation docs updated

* tests added

* Update large_scale_condensation.md

Fixed a typo, added a comment where I thought that the idea was unclear.

* Update large_scale_condensation.md

Tried to improve the paragraph about re-evaporation.

* docs updates

* longwave tests corrected

* docs math typo

* precipitation = rain + snow renamings

* consistent naming

* docs typo

---------

Co-authored-by: Milan Klöwer <milankloewer@gmx.de>